task5

+++ task3.py
@@ -1,3 +1,3 @@
 a = '38.2'
 b = 6.5
-print(float(a)*b)
+print(int(float(a)*b))
